LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-calcMode-discrete-path-expected.txt


FAIL Test animateMotion with calcMode="discrete" and path attribute assert_equals: point 1 (x) expected 0 but got 37.5
